[MOVE] SpaceMember 관련 DTO 패키지 따로 빼기
