+++ src/main/java/com/crewspace/api/dto/res/spaceMember/EnteredSpaceResponseDTO.java
@@ -1,4 +1,4 @@
-package com.crewspace.api.dto.res.space;
+package com.crewspace.api.dto.res.spaceMember;
 
 import com.crewspace.api.domain.spaceMember.SpaceMember;
 import java.util.List;
