+++ src/main/java/com/crewspace/api/service/SpaceMemberService.java
@@ -6,8 +6,8 @@ import com.crewspace.api.domain.member.Member;
 import com.crewspace.api.domain.member.MemberRepository;
 import com.crewspace.api.domain.spaceMember.SpaceMember;
 import com.crewspace.api.domain.spaceMember.SpaceMemberRepository;
-import com.crewspace.api.dto.req.space.EnteredSpaceRequestDTO;
-import com.crewspace.api.dto.res.space.EnteredSpaceResponseDTO;
+import com.crewspace.api.dto.req.spaceMember.EnteredSpaceRequestDTO;
+import com.crewspace.api.dto.res.spaceMember.EnteredSpaceResponseDTO;
 import com.crewspace.api.exception.CustomException;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
